requirements.txt
asgiref==3.3.1
cached-property==1.5.2
certifi==2020.12.5
chardet==4.0.0
charset-normalizer==1.3.4
Django==3.1.5
django-crispy-forms==1.10.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.13
djangorestframework==3.12.2
dragonmapper==0.2.6
hanzidentifier==1.0.2
idna==2.10
loguru==0.5.3
prettytable==2.0.0
pytz==2020.5
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
uWSGI==2.0.19.1
wcwidth==0.2.5
zhon==1.1.5
